schemas: move user notifications schemas from config to own file
